Create GetStates partial class
